+++ utf8.c
@@ -1,6 +1,6 @@
 #include "utf8.h"
 
-uint32_t utf8_cps(const char byte) {
+size_t utf8_cps(const char byte) {
     if (byte <= 0x7f) {
         return 1;
     } else if (byte <= 0xbf) {
@@ -16,11 +16,11 @@ uint32_t utf8_cps(const char byte) {
     return -1;
 }
 
-uint32_t utf8_length(const char* str) {
-    uint32_t length = 0;
-    uint32_t i = 0;
+size_t utf8_length(const char* str) {
+    size_t length = 0;
+    size_t i = 0;
     while (i < strlen(str)) {
-        uint32_t cps = utf8_cps(str[i]);
+        size_t cps = utf8_cps(str[i]);
         i += cps;
         length++;
     }
@@ -28,11 +28,11 @@ uint32_t utf8_length(const char* str) {
     return length;
 }
 
-uint32_t utf8_bytes(const char* str, uint32_t length) {
-    uint32_t num_bytes = 0;
-    uint32_t num_chars = 0;
+size_t utf8_bytes(const char* str, size_t length) {
+    size_t num_bytes = 0;
+    size_t num_chars = 0;
     while (num_chars < length) {
-        uint32_t cps = utf8_cps(str[num_bytes]);
+        size_t cps = utf8_cps(str[num_bytes]);
         num_bytes += cps;
         num_chars++;
     }
