+++ utf8.h
@@ -4,10 +4,10 @@
 #include <inttypes.h>
 #include <string.h>
 
-uint32_t utf8_cps(const char byte);
+size_t utf8_cps(const char byte);
 
-uint32_t utf8_length(const char* str);
+size_t utf8_length(const char* str);
 
-uint32_t utf8_bytes(const char* str, uint32_t length);
+size_t utf8_bytes(const char* str, size_t length);
 
 #endif
